operator/testing: cover DGDR workload pathways
