+++ deploy/operator/internal/controller/clusterenv_helpers_test.go
@@ -13,7 +13,65 @@ import (
 	"strings"
 
 	semver "github.com/Masterminds/semver/v3"
+	configv1alpha1 "github.com/ai-dynamo/dynamo/deploy/operator/api/config/v1alpha1"
+	nvidiacomv1alpha1 "github.com/ai-dynamo/dynamo/deploy/operator/api/v1alpha1"
+	nvidiacomv1beta1 "github.com/ai-dynamo/dynamo/deploy/operator/api/v1beta1"
+	grovev1alpha1 "github.com/ai-dynamo/grove/operator/api/core/v1alpha1"
+	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
+	istioclientsetscheme "istio.io/client-go/pkg/clientset/versioned/scheme"
+	admissionregistrationv1 "k8s.io/api/admissionregistration/v1"
+	appsv1 "k8s.io/api/apps/v1"
+	autoscalingv2 "k8s.io/api/autoscaling/v2"
+	batchv1 "k8s.io/api/batch/v1"
+	corev1 "k8s.io/api/core/v1"
+	discoveryv1 "k8s.io/api/discovery/v1"
+	networkingv1 "k8s.io/api/networking/v1"
+	rbacv1 "k8s.io/api/rbac/v1"
+	resourcev1 "k8s.io/api/resource/v1"
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/utils/ptr"
+	gaiev1 "sigs.k8s.io/gateway-api-inference-extension/api/v1"
+	lwsscheme "sigs.k8s.io/lws/client-go/clientset/versioned/scheme"
+	vcbatchv1alpha1 "volcano.sh/apis/pkg/apis/batch/v1alpha1"
+	volcanov1beta1 "volcano.sh/apis/pkg/apis/scheduling/v1beta1"
 )
+
+func newClusterTestScheme() *runtime.Scheme {
+	scheme := runtime.NewScheme()
+	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
+	utilruntime.Must(configv1alpha1.AddToScheme(scheme))
+	utilruntime.Must(nvidiacomv1alpha1.AddToScheme(scheme))
+	utilruntime.Must(nvidiacomv1beta1.AddToScheme(scheme))
+	utilruntime.Must(grovev1alpha1.AddToScheme(scheme))
+	utilruntime.Must(monitoringv1.AddToScheme(scheme))
+	utilruntime.Must(istioclientsetscheme.AddToScheme(scheme))
+	utilruntime.Must(lwsscheme.AddToScheme(scheme))
+	utilruntime.Must(gaiev1.Install(scheme))
+	utilruntime.Must(appsv1.AddToScheme(scheme))
+	utilruntime.Must(admissionregistrationv1.AddToScheme(scheme))
+	utilruntime.Must(autoscalingv2.AddToScheme(scheme))
+	utilruntime.Must(batchv1.AddToScheme(scheme))
+	utilruntime.Must(corev1.AddToScheme(scheme))
+	utilruntime.Must(discoveryv1.AddToScheme(scheme))
+	utilruntime.Must(networkingv1.AddToScheme(scheme))
+	utilruntime.Must(rbacv1.AddToScheme(scheme))
+	utilruntime.Must(resourcev1.AddToScheme(scheme))
+	utilruntime.Must(apiextensionsv1.AddToScheme(scheme))
+	utilruntime.Must(volcanov1beta1.AddToScheme(scheme))
+	utilruntime.Must(vcbatchv1alpha1.AddToScheme(scheme))
+	return scheme
+}
+
+func clusterTestRestrictedConfig(namespace string) *configv1alpha1.OperatorConfiguration {
+	config := &configv1alpha1.OperatorConfiguration{}
+	configv1alpha1.SetDefaultsOperatorConfiguration(config)
+	config.Namespace.Restricted = namespace
+	config.GPU.DiscoveryEnabled = ptr.To(false)
+	return config
+}
 
 func clusterTestOperatorVersion() (string, error) {
 	output, err := exec.Command(
