+++ deploy/operator/internal/controller/testdata/dgdr/grove/input.yaml
