+++ deploy/operator/internal/controller/testdata/dgdr/grove/output.yaml
